enqueuer/requirements.txt
--allow-all-external
python-twitter
